JUNE 5, 2016 CANSOLO FLIGHT SOFTWARE [WIP]
General flight software is complete. List of completed features below:

Features:
	State transition logic
	packet formation
	packet transmission
	saving state information
	command handling logic

TODO:
	fine tune state transition points
	look into pitot tube data handling
	investigate camera interrupts
	maintaining state information through reset

Plan to maintain state information by having extra logic in the init
section of code that checks to see if a flightdata.csv file already exists
on the SD card. If so then read necessary state information from there and
continue.

 On branch system_testing
	modified:   lib/BMP180/BMP180.cpp
	modified:   lib/TMP102/TMP102.cpp
	modified:   lib/TMP102/TMP102.hpp
	modified:   tests/system/system_test.cpp
	modified:   tests/system/system_test.hpp

+++ lib/BMP180/BMP180.cpp
@@ -131,7 +131,7 @@ float BMP180::getAltitude(float* p)
 {
     // formula for conversion directly from documentation
     float altitude = (*p / 1013.25);
-    altitude = pow(altitude, (1 / 5.255));
+    altitude = pow(altitude, (float)(1 / 5.255));
     altitude = 44330 * (1 - altitude);
     return altitude;
 }
